nvme: move args into separate header
Untaggle the global header a little by moving global nvme arguments
related defition to its own header. This prepares the stage for the
global ini config parser.
