+++ args.c
@@ -1,0 +1,16 @@
+// SPDX-License-Identifier: GPL-2.0-or-later
+/*
+ * This file is part of nvme-cli.
+ * Copyright (c) 2026 SUSE Software Solutions
+ *
+ * Authors: Daniel Wagner <dwagner@suse.de>
+ */
+
+#include "args.h"
+
+struct nvme_args nvme_args = {
+	.output_format = "normal",
+	.output_format_ver = 2,
+	.timeout = 0,
+	.supported_output_formats = DEFAULT_OUTPUT_FORMATS,
+};
