+++ nvme.c
@@ -242,13 +242,6 @@ static const char *pmrmscl = "PMRMSCL=0xe14 register offset";
 static const char *pmrmscu = "PMRMSCU=0xe18 register offset";
 static const char *ish = "Ignore Shutdown (for NVMe-MI command)";
 
-struct nvme_args nvme_args = {
-	.output_format = "normal",
-	.output_format_ver = 2,
-	.timeout = NVME_DEFAULT_IOCTL_TIMEOUT,
-	.supported_output_formats = DEFAULT_OUTPUT_FORMATS,
-};
-
 static void *mmap_registers(struct libnvme_transport_handle *hdl, bool writable);
 static int munmap_registers(void *addr);
